+++ rust/src/domains/polars/frame/mod.rs
@@ -381,46 +381,4 @@ impl Margin {
 }
 
 #[cfg(test)]
-mod test_lazyframe {
-    use super::*;
-    use crate::domains::AtomDomain;
-
-    #[test]
-    fn test_frame_new() -> Fallible<()> {
-        let lf_domain = LazyFrameDomain::new(vec![
-            SeriesDomain::new("A", AtomDomain::<i32>::default()),
-            SeriesDomain::new("B", AtomDomain::<f64>::default()),
-        ])?;
-
-        let lf = df!("A" => &[3, 4, 5], "B" => &[1., 3., 7.])?.lazy();
-
-        assert!(lf_domain.member(&lf)?);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_margin() -> Fallible<()> {
-        let lf_domain = LazyFrameDomain::new(vec![
-            SeriesDomain::new("A", AtomDomain::<i32>::default()),
-            SeriesDomain::new("B", AtomDomain::<String>::default()),
-        ])?
-        .with_margin(
-            &["A"],
-            Margin::new()
-                .with_max_partition_length(1)
-                .with_max_num_partitions(2),
-        )?;
-
-        let lf_exceed_partition_size = df!("A" => [1, 2, 2], "B" => ["1", "1", "2"])?.lazyframe();
-        assert!(!lf_domain.member(&lf_exceed_partition_size)?);
-
-        let lf_exceed_num_partitions = df!("A" => [1, 2, 3], "B" => ["1", "1", "1"])?.lazyframe();
-        assert!(!lf_domain.member(&lf_exceed_num_partitions)?);
-
-        let lf = df!("A" => [1, 2], "B" => ["1", "1"])?.lazyframe();
-        assert!(lf_domain.member(&lf)?);
-
-        Ok(())
-    }
-}
+mod test;
